Fix otel logger in example
