+++ opentelemetry-example/src/main/scala/zio/telemetry/opentelemetry/example/ProxyApp.scala
@@ -8,6 +8,7 @@ import zio.telemetry.opentelemetry.baggage.Baggage
 import zio.telemetry.opentelemetry.context.ContextStorage
 import zio.telemetry.opentelemetry.example.config.AppConfig
 import zio.telemetry.opentelemetry.example.http.{BackendClient, ProxyHttpApp, ProxyHttpServer}
+import zio.telemetry.opentelemetry.example.otel.{FluentbitLoggerProvider, FluentbitTracer}
 import zio.telemetry.opentelemetry.logging.Logging
 import zio.telemetry.opentelemetry.tracing.Tracing
 
@@ -17,11 +18,6 @@ object ProxyApp extends ZIOAppDefault {
 
   private val instrumentationScopeName = "zio.telemetry.opentelemetry.example.ProxyApp"
   private val resourceName             = "opentelemetry-example-proxy"
-
-  override val bootstrap: ZLayer[ZIOAppArgs, Throwable, Any] =
-    Runtime.removeDefaultLoggers >>>
-      (FluentbitLoggerProvider.live(resourceName) ++ ContextStorage.fiberRef) >>>
-      Logging.live(instrumentationScopeName)
 
   override def run: Task[ExitCode] =
     ZIO
@@ -35,7 +31,9 @@ object ProxyApp extends ZIOAppDefault {
         Tracing.live,
         Baggage.live(),
         ContextStorage.fiberRef,
-        FluentbitTracer.live(resourceName, instrumentationScopeName)
+        FluentbitTracer.live(resourceName, instrumentationScopeName),
+        FluentbitLoggerProvider.live(resourceName),
+        Logging.live(instrumentationScopeName)
       )
 
 }
