+++ opentelemetry-example/src/main/scala/zio/telemetry/opentelemetry/example/otel/SeqLoggerProvider.scala
@@ -1,4 +1,4 @@
-package zio.telemetry.opentelemetry.example
+package zio.telemetry.opentelemetry.example.otel
 
 import io.opentelemetry.api.common.Attributes
 import io.opentelemetry.api.logs.LoggerProvider
@@ -9,12 +9,21 @@ import io.opentelemetry.sdk.resources.Resource
 import io.opentelemetry.semconv.resource.attributes.ResourceAttributes
 import zio._
 
-object FluentbitLoggerProvider {
+/**
+ * https://datalust.co/seq
+ */
+object SeqLoggerProvider {
 
   def live(resourceName: String): TaskLayer[LoggerProvider] =
     ZLayer(
       for {
-        logRecordExporter  <- ZIO.succeed(OtlpHttpLogRecordExporter.builder().build())
+        logRecordExporter  <-
+          ZIO.succeed(
+            OtlpHttpLogRecordExporter
+              .builder()
+              .setEndpoint("http://localhost:5341/ingest/otlp/v1/logs")
+              .build()
+          )
         logRecordProcessor <- ZIO.succeed(SimpleLogRecordProcessor.create(logRecordExporter))
         loggerProvider     <-
           ZIO.attempt(
